requirements.txt
argparse==1.2.1
certifi==2017.4.17
chardet==3.0.4
docopt==0.6.2
idna==2.5
python-digitalocean==1.12
requests==2.18.1
urllib3==1.21.1
wsgiref==0.1.2
